feat(router): add <MyRoute> component and integrate it into MyRouterApp
- Created custom <MyRoute> component to conditionally render children based on current path
- Listens to popstate events to update internal state and re-render on navigation
- Refactored MyRouterApp to use declarative <MyRoute> components instead of manual routing logic
- Improves readability and separation of concerns in routing logic
