+++ routing-system/src/my-react-router/MyRouterApp.tsx
@@ -1,38 +1,10 @@
-import React, { useEffect, useState } from "react";
 import AccordionPage from "../pages/AccordionPage";
 import DropDownPage from "../pages/DropDownPage";
 import ButtonPage from "../pages/ButtonPage";
 import MyLink from "./MyLink";
+import MyRoute from "./MyRoute";
 
 const MyRouterApp = () => {
-  const [currentPath, setCurrentPath] = useState(window.location.pathname);
-
-  useEffect(() => {
-    const onLocationChange = () => {
-      setCurrentPath(window.location.pathname);
-    };
-
-    window.addEventListener("popstate", onLocationChange);
-
-    const cleanUp = () => {
-      window.removeEventListener("popstate", onLocationChange);
-    };
-
-    return cleanUp;
-  }, []);
-
-  // Routing logic (basic)
-  let ComponentToShow;
-  if (currentPath === "/accordion") {
-    ComponentToShow = <AccordionPage />;
-  } else if (currentPath === "/dropdown") {
-    ComponentToShow = <DropDownPage />;
-  } else if (currentPath === "/buttons") {
-    ComponentToShow = <ButtonPage />;
-  } else {
-    ComponentToShow = <div>Home Page</div>;
-  }
-
   return (
     <div>
       <nav style={{ marginBottom: "1rem" }}>
@@ -41,7 +13,19 @@ const MyRouterApp = () => {
         <MyLink to="/buttons">Buttons</MyLink>
       </nav>
       <hr />
-      {ComponentToShow}
+
+      <MyRoute path="/accordion">
+        <AccordionPage />
+      </MyRoute>
+      <MyRoute path="/dropdown">
+        <DropDownPage />
+      </MyRoute>
+      <MyRoute path="/buttons">
+        <ButtonPage />
+      </MyRoute>
+      <MyRoute path="/">
+        <div>Home Page</div>
+      </MyRoute>
     </div>
   );
 };
